+++ frontend/package.json
@@ -11,11 +11,11 @@
     "test": "jest"
   },
   "dependencies": {
-    "@apollo/client": "^3.10.1",
+    "@apollo/client": "^3.10.2",
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
-    "@mui/icons-material": "^5.15.15",
-    "@mui/material": "^5.15.15",
+    "@mui/icons-material": "^5.15.16",
+    "@mui/material": "^5.15.16",
     "axios": "^1.6.8",
     "graphql": "^16.8.1",
     "next": "^14.2.3",
@@ -27,11 +27,11 @@
     "ts-node": "^10.9.2"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^6.4.2",
-    "@testing-library/react": "^15.0.5",
+    "@testing-library/jest-dom": "^6.4.5",
+    "@testing-library/react": "^15.0.6",
     "@types/jest": "^29.5.12",
-    "@types/node": "20.12.7",
-    "@types/react": "18.3.0",
+    "@types/node": "20.12.8",
+    "@types/react": "18.3.1",
     "@types/react-dom": "18.3.0",
     "@types/react-google-recaptcha": "^2.1.9",
     "eslint": "^8.57.0",
@@ -43,7 +43,7 @@
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "prettier": "^3.2.5",
-    "sass": "^1.75.0",
+    "sass": "^1.76.0",
     "typescript": "^5.4.5"
   }
 }
