+++ frontend/src/pages/_app.tsx
@@ -1,21 +1,13 @@
 import "@/styles/index.scss";
-import {
-  ApolloClient,
-  ApolloProvider,
-  HttpLink,
-  InMemoryCache,
-  useQuery,
-} from "@apollo/client";
+import { ApolloProvider } from "@apollo/client";
 import type { AppProps } from "next/app";
 import dynamic from "next/dynamic";
-import { createTheme, ThemeProvider } from "@mui/material/styles";
-import { Box, CircularProgress, CssBaseline, Typography } from "@mui/material";
-import { queryMeContext } from "@/components/graphql/Users";
-import { useEffect } from "react";
-import { useRouter } from "next/router";
-import { UserContextTypes } from "@/types/UserTypes";
+import { ThemeProvider, createTheme } from "@mui/material/styles";
+import { client } from "@/config/apollo-client";
+import { useAuth } from "@/config/userAuth";
+import { CssBaseline } from "@mui/material";
 import Header from "@/components/appBar/AppBar";
-import { API_URL } from "@/api/configApi";
+import { LoadingApp } from "@/styles/LoadingApp";
 import { VariablesColors } from "@/styles/Variables.colors";
 
 const colors = new VariablesColors();
@@ -40,90 +32,27 @@ const theme = createTheme({
       fontFamily: ["Poppins-Medium", "sans-serif"].join(","),
     },
   },
-
   palette: {
     mode: "light",
-    background: {
-      default: colorWhite,
-    },
+    background: { default: colorWhite },
     primary: {
       main: colorOrange,
       light: colorLightOrange,
       dark: colorDarkOrange,
     },
-    secondary: {
-      main: "#343a40",
-      light: "#5C6166",
-      dark: "#24282C",
-    },
+    secondary: { main: "#343a40", light: "#5C6166", dark: "#24282C" },
   },
   components: {
-    MuiPaper: {
-      styleOverrides: {
-        root: {
-          backgroundColor: "#F8F8F8",
-        },
-      },
-    },
+    MuiPaper: { styleOverrides: { root: { backgroundColor: "#F8F8F8" } } },
   },
 });
-const client = new ApolloClient({
-  link: new HttpLink({
-    uri: API_URL || "http://localhost:5000/",
-    credentials: "include",
-  }),
-  cache: new InMemoryCache(),
-});
+const privatePages = ["/account", "/annonces/new"];
 
-const privatePages = ["/compte", "/annonces/new"];
-
-function AuthProvider({ children }: { children: React.ReactNode }) {
-  const { loading, error, refetch } = useQuery<{
-    item: UserContextTypes;
-  }>(queryMeContext);
-  const router = useRouter();
-  useEffect(() => {
-    const handleRouteChange = () => {
-      if (privatePages.includes(router.pathname)) {
-        refetch();
-      }
-    };
-
-    router.events.on("routeChangeComplete", handleRouteChange);
-
-    return () => {
-      router.events.off("routeChangeComplete", handleRouteChange);
-    };
-  }, [router, refetch]);
-
-  useEffect(() => {
-    if (privatePages.includes(router.pathname) && error) {
-      router.replace("/signin");
-    }
-  }, [router, error]);
-
-  if (loading) {
-    return (
-      <Box
-        sx={{
-          display: "flex",
-          flexDirection: "column",
-          justifyContent: "center",
-          alignItems: "center",
-          height: "100vh",
-          width: "100vw",
-        }}
-      >
-        <CircularProgress size={120} />
-        <Typography variant="h4" gutterBottom>
-          Chargement...
-        </Typography>
-      </Box>
-    );
-  }
-
+const AuthProvider = ({ children }: { children: React.ReactNode }) => {
+  const { loading, error } = useAuth(privatePages);
+  if (loading) return <LoadingApp />;
   return children;
-}
+};
 
 function App({ Component, pageProps }: AppProps) {
   return (
